Moved Exceptions to common.py file to avoid upcoming circular depencies
